ci(perf_website_sale): prevent us from getting a green tick in ci, we should probably ensure that UNACCENT is present in ci, but this will have to do for the mment

+++ perf_website_sale/models/product_pricelist.py
@@ -1,4 +1,9 @@
+import logging
+
 from odoo import models
+from odoo.modules.db import FunctionStatus
+
+_logger = logging.getLogger(__name__)
 
 
 class ProductPricelistItem(models.Model):
@@ -6,6 +11,12 @@ class ProductPricelistItem(models.Model):
 
     def init(self):
         res = super().init()
+        if self.env.registry.has_unaccent != FunctionStatus.INDEXABLE:
+            _logger.critical(
+                "UNACCENT is not present and INDEXABLE, cannot apply perf_website_sale"
+            )
+            return res
+
         indexes_to_remove = [
             "DROP INDEX IF EXISTS test_product_pricelist_item_product_id",
             "DROP INDEX IF EXISTS test_product_pricelist_item_product_tmpl_id",
