+++ perf_website_sale/models/product_template.py
@@ -1,5 +1,10 @@
+import logging
+
 from odoo import api, fields, models
+from odoo.modules.db import FunctionStatus
 from odoo.tools.sql import SQL, column_exists, create_column
+
+_logger = logging.getLogger(__name__)
 
 # A delimiter that users aren't likely to search for in product codes.
 # XXX: \u241e is the unicode record separator character. Do not waste time over
@@ -22,6 +27,12 @@ class ProductTemplate(models.Model):
 
     def init(self):
         res = super().init()
+        if self.env.registry.has_unaccent != FunctionStatus.INDEXABLE:
+            _logger.critical(
+                "UNACCENT is not present and INDEXABLE, cannot apply perf_website_sale"
+            )
+            return res
+
         indexes_to_remove = [
             "DROP INDEX IF EXISTS test_product_template_sale_ok",
             "DROP INDEX IF EXISTS test_product_template_active",
